Revert "Enables dead code elimination algorithm: Three Shaking (feature of webpack 2 and babel 6)"
This reverts commit 5450c68f36ecd3673126321009795e35a321d92b.

Reason: ng-annotate-loader needs babel to transpile ES6 code to valid ES5 and
this is being achieved by the used of the option "modules": false on the babel 2015 preset.

There is the ng-annotate-webpack-plugin that works with Webpack2/Babel6 Tree Shaking. However
there is no option to specify the chunk/files where annotation should be applied, which makes
the bundling process really slow (prefer not to use it).